Collapse file storage code into Cache class
