+++ spec/integration/core/cache_spec.rb
@@ -3,11 +3,11 @@
 RSpec.describe Spandx::Core::Cache do
   RSpec.shared_examples 'each data file' do |package_manager, key|
     describe "#licenses_for (#{package_manager})" do
-      subject { described_class.new(package_manager, db: Spandx.git[key]) }
+      subject { described_class.new(package_manager, root: Spandx.git[key].path) }
 
       (0x00..0xFF).map { |x| x.to_s(16).upcase.rjust(2, '0').downcase }.each do |hex|
         context hex do
-          let(:path) { subject.db.expand_path(".index/#{hex}/#{package_manager}") }
+          let(:path) { subject.expand_path(".index/#{hex}/#{package_manager}") }
 
           it "is able to find all packages in the #{package_manager} index" do
             CSV.foreach(path) do |row|
