+++ spec/unit/dotnet/index_spec.rb
@@ -1,8 +1,9 @@
 # frozen_string_literal: true
 
 RSpec.describe Spandx::Dotnet::Index do
-  subject { described_class.new(directory: directory) }
+  subject { described_class.new(directory: directory, gateway: gateway) }
 
+  let(:gateway) { instance_double(Spandx::Dotnet::NugetGateway) }
   let(:directory) { Dir.mktmpdir('spandx') }
 
   after do
@@ -10,17 +11,14 @@ RSpec.describe Spandx::Dotnet::Index do
   end
 
   describe '#update!' do
-    let(:catalogue) { Spandx::Spdx::Catalogue.from_file(fixture_file('spdx/json/licenses.json')) }
-    let(:gateway) { instance_double(Spandx::Dotnet::NugetGateway) }
+    let(:item) { { 'id' => 'Polaroider', 'version' => '0.2.0', 'licenseExpression' => 'MIT' } }
 
     before do
-      allow(Spandx::Dotnet::NugetGateway).to receive(:new).and_return(gateway)
-      item = { 'id' => 'Polaroider', 'version' => '0.2.0', 'licenseExpression' => 'MIT' }
-      allow(gateway).to receive(:each).and_yield(item, 0)
+      allow(gateway).to receive(:each).and_yield(item)
 
-      subject.update!(catalogue: catalogue)
+      subject.update!(db: db)
     end
 
-    specify { expect(subject.licenses_for(name: 'Polaroider', version: '0.2.0')).to match_array(['MIT']) }
+    specify { expect(db.licenses_for('Polaroider', '0.2.0')).to match_array(['MIT']) }
   end
 end
